fix(spaces): Fixes performance in expanding spaces (#64)
Fixes a performance issue when expanding a colleciton of spaces
with information about their owners. Removed use of forkJoin in
fetching owners for a collection of spaces, due to difficulty in
sharing/cache the HTTP response or Observable<User> across
parallel streams. The fix involves obtaining a stream of distinct
owners, fetching the Users from the backend, and then populating
the spaces. Fixes #56